Split hdf5_io.py 	-> tools/hdf5_handler.py #generic function to handle hdf5 	-> tools/object_handler.py #only handle object 	-> io/hdf5_io.py #only handle hdf5 <=> smash model object
